requirements.txt
discord.py
psutil
matplotlib
numpy
speedtest-cli
aiohttp
